Quick error fixed. I had a freeze in my playthrough!
When I was going through my game today. (submited this morning at 5) I found a freeze. The only thing wrong was the UI was checked non active in the last level because I was tersting. But it froze the last level. This one works fine. The only change is the setActive checkbox on my UI game object.